+++ Web/vue2/src/router/index.ts
@@ -1,6 +1,5 @@
 import Vue from 'vue';
 import VueRouter from 'vue-router';
-import Home from '../views/home.vue';
 
 Vue.use(VueRouter);
 
@@ -8,12 +7,12 @@ const routes = [
   {
     path: '/',
     name: 'home',
-    component: Home,
+    component: () => import(/* webpackChunkName: "about" */ '../views/home.vue'),
   },
   {
     path: '/home',
     name: 'home',
-    component: Home,
+    component: () => import(/* webpackChunkName: "about" */ '../views/home.vue'),
   },
   {
     path: '/room',