+++ Web/vue2/src/utils/utils.ts
@@ -1,0 +1,40 @@
+
+/**
+ * 判断字符串是否是数字
+ * @returns boolean
+ */
+export function checkNumber(roomId: string) {
+  const reg = /^\d+$/;
+  return reg.test(roomId);
+}
+
+/**
+ * 从 window.location.href 中获取指定key的value
+ * @param {*} key 要获取的 key
+ * @returns window.location.href 中指定key对应的value
+ * @example
+ * const value = getUrlParam(key);
+ */
+export function getUrlParam(key: string) {
+  const url = window?.location.href.replace(/^[^?]*\?/, '');
+  const regexp = new RegExp(`(^|&)${key}=([^&#]*)(&|$|)`, 'i');
+  const paramMatch = url?.match(regexp);
+
+  return paramMatch ? paramMatch[2] : null;
+}
+
+/**
+ * 获取语言
+ * @returns language
+ */
+export function getLanguage() {
+  let language = getUrlParam('lang') || localStorage.getItem('tuiRoom-language') || navigator.language || 'zh';
+  language = language.replace(/_/, '-').toLowerCase();
+
+  if (language === 'zh-cn' || language === 'zh') {
+    language = 'zh-CN';
+  } else if (language === 'en' || language === 'en-us' || language === 'en-GB') {
+    language = 'en-US';
+  }
+  return language;
+}